_legacy/backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0

# Redis
redis[hiredis]==5.0.1

# WebSocket
websockets==12.0
python-socketio==5.10.0

# Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx==0.25.2
structlog==23.2.0
aiofiles==23.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
aiosqlite==0.19.0
black==23.12.0
isort==5.12.0
pylint==3.0.3
mypy==1.7.1

# Code Quality & Security Tools (for evaluation framework)
radon==6.0.1
bandit==1.7.5
flake8==7.0.0

# Monitoring
prometheus-client==0.19.0

# Protocol Buffers (ensure compatibility with opentelemetry)
protobuf>=3.19,<5.0
